fix(#73): 유저 ID 토큰에서 추출해 공유 포트폴리오 댓글 CRUD와 연결

+++ portfolio-service/src/main/java/com/pda/portfolioservice/service/PortfolioServiceImpl.java
@@ -2,24 +2,29 @@ package com.pda.portfolioservice.service;
 
 import com.pda.portfolioservice.dto.request.SharePortfolioCommentRequestDTO;
 import com.pda.portfolioservice.dto.response.MyPortfolioTitleResponseDTO;
+import com.pda.portfolioservice.dto.response.NicknameResponseDTO;
 import com.pda.portfolioservice.dto.response.ShareMyPortfolioResponseDTO;
 import com.pda.portfolioservice.dto.response.SharePortfolioCommentResponseDTO;
 import com.pda.portfolioservice.entity.MyPortfolio;
 import com.pda.portfolioservice.entity.SharePortfolio;
 import com.pda.portfolioservice.entity.SharePortfolioComment;
+import com.pda.portfolioservice.feign.UserServiceClient;
 import com.pda.portfolioservice.model.Portfolio;
 import com.pda.portfolioservice.repository.MyPortfolioRepository;
 import com.pda.portfolioservice.repository.PortfolioRepository;
 import com.pda.portfolioservice.repository.SharePortfolioCommentRepository;
 import com.pda.portfolioservice.repository.SharePortfolioRepository;
+import com.pda.utilservice.jwt.JWTUtil;
 import com.pda.utilservice.response.code.resultCode.ErrorStatus;
 import com.pda.utilservice.response.exception.handler.PortfolioHandler;
 import lombok.RequiredArgsConstructor;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -30,6 +35,9 @@ public class PortfolioServiceImpl implements PortfolioService {
     private final SharePortfolioRepository sharePortfolioRepository;
     private final SharePortfolioCommentRepository sharePortfolioCommentRepository;
     private final PortfolioRepository portfolioRepository;
+
+    private final UserServiceClient userServiceClient;
+    private final Environment environment;
 
     // 포트폴리오 저장 (중복 검사 후 저장)
     @Override
@@ -101,21 +109,28 @@ public class PortfolioServiceImpl implements PortfolioService {
     }
 
     @Override
-    public void addComment(Long sharePortfolioId, SharePortfolioCommentRequestDTO requestDTO) {
+    public void addComment(Long sharePortfolioId, SharePortfolioCommentRequestDTO requestDTO, String token) {
+        System.out.println("token = " + token);
+        JWTUtil jwtUtil = new JWTUtil(Objects.requireNonNull(environment.getProperty("spring.jwt.secret")));
+        String userId = jwtUtil.getBearerUserId(token);
+
+        System.out.println(userId);
+
         SharePortfolio sharePortfolio = sharePortfolioRepository.findById(sharePortfolioId)
                 .orElseThrow(() -> new PortfolioHandler(ErrorStatus.PORTFOLIO_NOT_FOUND));
 
         SharePortfolioComment comment = SharePortfolioComment.builder()
                         .sharePortfolio(sharePortfolio)
-                                .content(requestDTO.getContent())
-                                        .build();
+                        .userId(userId)
+                        .content(requestDTO.getContent())
+                        .build();
 
         sharePortfolioCommentRepository.save(comment);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public SharePortfolioCommentResponseDTO getComments(Long sharePortfolioId) {
+    public SharePortfolioCommentResponseDTO getComments(Long sharePortfolioId, String userId) {
         List<SharePortfolioComment> comments = sharePortfolioCommentRepository.findBysharePortfolio_SharePortfolioId(sharePortfolioId);
 
         if (comments.isEmpty()) {
@@ -125,12 +140,19 @@ public class PortfolioServiceImpl implements PortfolioService {
                     .build();
         }
 
-        return SharePortfolioCommentResponseDTO.toDTO(comments);
+        NicknameResponseDTO nicknameResponseDTO = userServiceClient.getNickname(userId);
+        String nickname = nicknameResponseDTO.getNickname();
+
+
+        return SharePortfolioCommentResponseDTO.toDTO(comments, nickname);
 
     }
 
     @Override
-    public void updateComment(Long sharePortfolioId, Long commentId, SharePortfolioCommentRequestDTO requestDTO) {
+    public void updateComment(Long sharePortfolioId, Long commentId, SharePortfolioCommentRequestDTO requestDTO, String token) {
+        JWTUtil jwtUtil = new JWTUtil(Objects.requireNonNull(environment.getProperty("spring.jwt.secret")));
+        String userId = jwtUtil.getBearerUserId(token);
+
         SharePortfolioComment comment = sharePortfolioCommentRepository.findById(commentId)
                 .orElseThrow(() -> new PortfolioHandler(ErrorStatus.PORTFOLIO_COMMENT_NOT_FOUND));
 
@@ -138,17 +160,26 @@ public class PortfolioServiceImpl implements PortfolioService {
             throw new PortfolioHandler(ErrorStatus.PORTFOLIO_COMMENT_NOT_INCLUDED);
         }
 
+        if (!comment.getUserId().equals(userId)) {
+            throw new PortfolioHandler(ErrorStatus.UNAUTHORIZED);
+        }
+
         comment.setContent(requestDTO.getContent());
+        sharePortfolioCommentRepository.save(comment);
     }
 
     @Override
-    public void deleteComment(Long sharePortfolioId, Long commentId) {
+    public void deleteComment(Long sharePortfolioId, Long commentId, String token) {
+        JWTUtil jwtUtil = new JWTUtil(Objects.requireNonNull(environment.getProperty("spring.jwt.secret")));
+        String userId = jwtUtil.getBearerUserId(token);
+
         SharePortfolioComment comment = sharePortfolioCommentRepository.findById(commentId)
                 .orElseThrow(() -> new PortfolioHandler(ErrorStatus.PORTFOLIO_COMMENT_NOT_FOUND));
 
+        if (!comment.getUserId().equals(userId)) {
+            throw new PortfolioHandler(ErrorStatus.UNAUTHORIZED);
+        }
+
         sharePortfolioCommentRepository.deleteById(commentId);
-
     }
-
-
 }
